Major Changes
When viewing the shop, if an item is out of stock it now renders a different card to display in large red text "Out of Stock".

When adding items to the cart they now display how many items are in the cart in the NavBar.

To add items to the cart you need to logged in. If you are not logged in it will redirect you to the Login Page.

If not logged in, the Cart Page now asks you to login to view your cart.

If you are logged in and your cart is empty it says so.

There is also now a bar in the cart that display total items and total cost as well as a check out button.

On the Cart Page it is now possible to remove items from your cart.

Checkout button works to empty cart and update DB appropriately.

Added the main content to About Page,  Added a Map to View the Store "location".